requirements.txt
beautifulsoup4==4.5.1
boto==2.42.0
dj-static==0.0.6
Django==1.9.8
django-appconf==1.0.2
django-boto==0.3.9
django-compressor==2.1
django-debug-toolbar==1.5
django-extra-views==0.8.0
django-filter==0.13.0
django-htmlmin==0.9.1
django-storages==1.5.0
djangorestframework==3.4.3
gunicorn==19.6.0
html5lib==0.999999999
Pillow==3.3.0
pyjade==4.0.0
python-dateutil==2.5.3
querystring-parser==1.2.3
rcssmin==1.0.6
requests==2.10.0
rjsmin==1.0.12
six==1.10.0
sqlparse==0.2.0
static3==0.7.0
webencodings==0.5
